refactor(storage): extract AuditEventStoreMixin from governance (SQLite + ABC)
First extraction of the _governance.py decomposition. Move the AuditEventStore
bucket — append_audit_event, list_audit_events, gc_governance_retention, and the
Audit-owned private _append_audit_event_with_cursor — out of GovernanceMixin into
a new governance/_audit.py AuditEventStoreMixin, verbatim (byte-identical bodies,
ERASE-guard and atomicity preserved).

The residual GovernanceMixin stays composed alongside AuditEventStoreMixin in the
sqlite_storage and storage_base base lists; it permanently holds shared infra
(conn, _lock, org_id, _deps()) and the module-level helpers (_json_dumps,
_row_to_audit_event), which _audit.py imports rather than duplicates. The
cross-bucket purge/barrier completion methods reach the moved
_append_audit_event_with_cursor via MRO (typed stub added on the residual).
