+++ reflexio/server/services/storage/sqlite_storage/_governance.py
@@ -3,6 +3,7 @@ from __future__ import annotations
 import json
 import sqlite3
 import threading
+from collections.abc import Callable
 from typing import Any, Literal, Protocol, cast
 
 from reflexio.models.api_schema.domain import AgentPlaybook, AgentPlaybookSourceWindow
@@ -14,7 +15,6 @@ from reflexio.models.api_schema.domain.governance import (
     SubjectBarrierStatus,
     SubjectWriteBarrier,
 )
-from reflexio.models.config_schema import GovernanceRetentionConfig
 from reflexio.server.services.governance.config import (
     get_governance_ref_secret,
     governance_subject_ref,
@@ -360,6 +360,12 @@ class SQLiteGovernanceMixin:
     _lock: threading.RLock
     org_id: str
 
+    # Provided via MRO by the co-composed AuditEventStoreMixin (audit bucket);
+    # reached here by the cross-bucket purge / barrier completion methods.
+    _append_audit_event_with_cursor: Callable[
+        [sqlite3.Connection | sqlite3.Cursor, AuditEvent], bool
+    ]
+
     def _deps(self) -> _SQLiteGovernanceDeps:
         return cast(_SQLiteGovernanceDeps, self)
 
@@ -682,31 +688,6 @@ class SQLiteGovernanceMixin:
             )
         )
 
-    def _append_audit_event_with_cursor(
-        self, cur: sqlite3.Connection | sqlite3.Cursor, event: AuditEvent
-    ) -> bool:
-        inserted = cur.execute(
-            """INSERT OR IGNORE INTO audit_events (
-                   org_id, actor_type, actor_ref, operation, entity_type, entity_id,
-                   subject_ref, request_ref, idempotency_key, status, detail, created_at
-               ) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)""",
-            (
-                event.org_id,
-                event.actor_type,
-                event.actor_ref,
-                event.operation,
-                event.entity_type,
-                event.entity_id,
-                event.subject_ref,
-                event.request_ref,
-                event.idempotency_key,
-                event.status,
-                _json_dumps(event.detail),
-                event.created_at,
-            ),
-        )
-        return inserted.rowcount > 0
-
     def _record_purge_target_locked(
         self,
         *,
@@ -797,35 +778,6 @@ class SQLiteGovernanceMixin:
                WHERE purge_id = ? AND org_id = ?""",
             (status, now, purge_id, self.org_id),
         )
-
-    def append_audit_event(self, event: AuditEvent) -> bool:
-        if _is_successful_erase_event(event):
-            raise ValueError(
-                "Successful ERASE audit rows may only be written by "
-                "complete_purge_operation_with_audit()"
-            )
-        if event.org_id != self.org_id:
-            raise ValueError("Audit event org_id must match storage org_id")
-        event = _canonicalize_audit_event_for_persistence(event)
-        with self._lock:
-            inserted = self._append_audit_event_with_cursor(self.conn, event)
-            self.conn.commit()
-            return inserted
-
-    def list_audit_events(
-        self, subject_ref: str | None = None, *, org_id: str | None = None
-    ) -> list[AuditEvent]:
-        deps = self._deps()
-        if org_id is not None and org_id != self.org_id:
-            raise ValueError("Audit event org_id must match storage org_id")
-        sql = "SELECT * FROM audit_events WHERE org_id = ?"
-        params: list[Any] = [self.org_id]
-        if subject_ref is not None:
-            sql += " AND subject_ref = ?"
-            params.append(subject_ref)
-        sql += " ORDER BY created_at ASC, event_id ASC"
-        rows = deps._fetchall(sql, params)
-        return [_row_to_audit_event(row) for row in rows]
 
     def _purge_governance_entity_content_locked(
         self,
@@ -1939,27 +1891,3 @@ class SQLiteGovernanceMixin:
         if row is None:
             raise ValueError(f"Purge operation {purge_id!r} not found")
         return _row_to_purge_operation(row)
-
-    def gc_governance_retention(self, *, config: GovernanceRetentionConfig) -> int:
-        if not config.audit_events_retention_enabled:
-            return 0
-        cutoff_epoch = _epoch_now() - config.audit_events_retention_days * 24 * 60 * 60
-        with self._lock:
-            cur = self.conn.execute(
-                """DELETE FROM audit_events
-                   WHERE event_id IN (
-                       SELECT event_id
-                       FROM audit_events
-                       WHERE org_id = ? AND created_at < ?
-                       ORDER BY created_at ASC, event_id ASC
-                       LIMIT ?
-                   )""",
-                (
-                    self.org_id,
-                    cutoff_epoch,
-                    config.audit_events_delete_batch_limit,
-                ),
-            )
-            deleted = int(cur.rowcount or 0)
-            self.conn.commit()
-        return deleted
