+++ reflexio/server/services/storage/sqlite_storage/governance/__init__.py
@@ -1,0 +1,5 @@
+from ._audit import AuditEventStoreMixin
+
+__all__ = [
+    "AuditEventStoreMixin",
+]
